requirements.txt
Django==3.1.4
django-phonenumber-field==5.0.0
django-rest-knox==4.1.0
djangorestframework==3.12.2

